Custom fluent handler API, Episode 20

+++ Passenger.Infrastructure/Handlers/Drivers/DeleteDriverRouteHandler.cs
@@ -1,5 +1,6 @@
 ﻿using Passenger.Infrastructure.Commands;
 using Passenger.Infrastructure.Commands.Drivers;
+using Passenger.Infrastructure.Exceptions;
 using Passenger.Infrastructure.Services;
 
 namespace Passenger.Infrastructure.Handlers.Drivers;
@@ -15,6 +16,11 @@ public class DeleteDriverRouteHandler : ICommandHandler<DeleteDriverRoute>
 
     public async Task HandleAsync(DeleteDriverRoute command)
     {
+        if (command.Name is null)
+        {
+            throw new ServiceException(Exceptions.ErrorCodes.NullCommandParameter, "Name parameter cannot be null");
+        }
+
         await _driverRouteService.DeleteAsync(command.UserId, command.Name);
     }
 }
